deletin items from custom todo list

+++ app.js
@@ -7,34 +7,34 @@ const app = express();
 app.use(bodyParser.urlencoded({extended:true}));
 app.use(express.static(__dirname + '/public'));
 
-app.set('view engine', 'ejs')
+app.set('view engine', 'ejs');
 
-mongoose.connect('mongodb://localhost:27017/todolistDB')
+mongoose.connect('mongodb://localhost:27017/todolistDB');
 
 const itemSchema = {
   name: String
-}
+};
 
-const Item = mongoose.model('Item', itemSchema)
+const Item = mongoose.model('Item', itemSchema);
 
 const item1 = new Item({
   name : 'Welcome to your todolist!'
-})
+});
 
 const item2 = new Item({
   name : 'Hit the + button to add a new task'
-})
+});
 
 const item3 = new Item({
   name : '<-- Hit this to cross out an item'
-})
+});
 
 const defaultItems = [item1, item2, item3];
 
 const listSchema = {
   name: String,
   items: [itemSchema]
-}
+};
 
 const List = mongoose.model('List', listSchema)
 
@@ -54,7 +54,7 @@ app.get('/:listName', (req, res) =>{
         res.render('list', {listTitle: foundList.name, newListItems:foundList.items})
       }
     }
-  })
+  });
 })
 
 app.get("/", (req, res) => {
@@ -77,47 +77,57 @@ app.get("/", (req, res) => {
       });
     }
 
-  })
+  });
 
 });
 
 app.post('/', (req, res) => {
 
   const itemName = req.body.newItem;
+  const listName = req.body.btn;
 
   const item = new Item({
     name: itemName
-  })
-  item.save()
-  res.redirect('/')
+  });
 
-  // if (req.body.list == 'Work'){
-  //   workItems.push(item)
-  //   res.redirect('/work')
-  // } else {
-  //   items.push(item);
-  // }
+  if(listName === 'Today'){
+    item.save()
+    res.redirect('/')
+  } else {
+    List.findOne({name: listName}, (err, foundList) => {
+      if(!err){
+        foundList.items.push(item);
+        foundList.save();
+        res.redirect('/'+ listName);
+      } else {
+        console.log(err);
+      }
+    });
+  }
 
 });
 
 app.post('/delete', (req, res) => {
   const checkedItemId = req.body.checkbox;
+  const listName = req.body.listName;
 
-  Item.findByIdAndRemove(checkedItemId, err =>{
-    if (!err){
-      console.log('Removed Item');
-      res.redirect('/')
-    }
-  })
+  if(listName === 'Today'){
+    Item.findByIdAndRemove(checkedItemId, err =>{
+      if (!err){
+        console.log('Removed Item');
+        res.redirect('/');
+      }
+    });
+  } else {
+    List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}}, (err, foundList) =>{
+      if(!err){
+        res.redirect('/'+ listName);
+      }
+    });
+  }
 
-})
-
-app.post('/work', (req, res) => {
-  let item = req.body.newItem;
-  workItems.push(item);
-  res.redirect('/work')
-})
+});
 
 app.listen(3000, () => {
   console.log("Server running on port 3000");
-})
+});
